+++ nodes/EvolutionApi/execute/chatwootRequest.ts
@@ -1,0 +1,18 @@
+import { IExecuteFunctions, IRequestOptions } from 'n8n-workflow';
+
+export async function chatwootRequest(ef: IExecuteFunctions, options: IRequestOptions) {
+	const credentials = await ef.getCredentials('httpbinApi');
+	const serverUrl = credentials['server-url'];
+	const apiKey = credentials.apikey;
+
+	const requestOptions: IRequestOptions = {
+		...options,
+		headers: {
+			apikey: apiKey,
+			...(options.headers || {}),
+		},
+		uri: `${serverUrl}${options.uri}`,
+	};
+
+	return await ef.helpers.request(requestOptions);
+}
