+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "n8n-nodes-evolution-api",
-  "version": "0.2.72",
+  "version": "0.2.73",
   "description": "Api não oficial do WhatsApp",
   "keywords": [
     "n8n-community-node-package"
@@ -36,10 +36,10 @@
   "n8n": {
     "n8nNodesApiVersion": 1,
     "credentials": [
-      "dist/credentials/HttpBinApi.credentials.js"
+      "dist/credentials/EvolutionApi.credentials.js"
     ],
     "nodes": [
-      "dist/nodes/HttpBin/HttpBin.node.js"
+      "dist/nodes/EvolutionApi/EvolutionApi.node.js"
     ]
   },
   "devDependencies": {
